plugins/caesar_cipher/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
nats-py==2.14.0
